changed split len to size t

+++ headers/libft.h
@@ -6,7 +6,7 @@
 /*   By: brfialho <brfialho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/12 16:45:30 by brfialho          #+#    #+#             */
-/*   Updated: 2025/10/22 20:37:21 by brfialho         ###   ########.fr       */
+/*   Updated: 2025/10/23 12:24:37 by brfialho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -175,7 +175,7 @@ char	**ft_split(char const *s, char c)
 		__attribute__((nonnull(1)));
 void	ft_split_free(char **split)
 		__attribute__((nonnull(1)));
-int		ft_split_len(char **split);
+size_t		ft_split_len(char **split);
 void	ft_split_print(char **split)
 		__attribute__((nonnull(1)));
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
