+++ srcs/str/ft_split_len.c
@@ -6,13 +6,15 @@
 /*   By: brfialho <brfialho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/26 17:26:02 by brfialho          #+#    #+#             */
-/*   Updated: 2025/09/26 17:26:23 by brfialho         ###   ########.fr       */
+/*   Updated: 2025/10/23 12:18:05 by brfialho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_split_len(char **split)
+#include "libft.h"
+
+size_t	ft_split_len(char **split)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	if (!split)
